make trait out of AbstractUser
